В опциях удалены флаги логирования запросов в трассы и логи, логирование ведется в лог по умолчанию.
